move magic command line tools to its own file
