+++ src/pestuary.py
@@ -1,7 +1,5 @@
 import inspect
-from functools import wraps
 import click
-import jsonpickle
 import os
 import sys
 import time
@@ -10,7 +8,6 @@ import requests
 import swagger_client
 from swagger_client.rest import ApiException
 from pprint import pprint
-
 
 #ESTUARY_URL='http://localhost:3004'
 ESTUARY_URL='https://api.estuary.tech'
@@ -50,58 +47,29 @@ try:
 except ApiException as e:
     print("Exception when calling UserApi->user_api_keys_get: %s\n" % e)
 
-@click.group()
-def cli():
-    pass
-
-@cli.result_callback()
-def process_result(result):
-    print(json.dumps(result, default=lambda x: x.to_dict()))
-    return result
 
 
 
-#wrote magic to automate some things that clck doesnt do, i didnt want to have to specify each argument to click for each method when it can be done automatically using decorators
-def magic(fn):
-    if __name__ != "__main__":
-        return fn
-    argspec = inspect.getargspec(fn)
-    arguments = argspec.args
-    defaults = argspec.defaults
-    if not arguments:
-        arguments = []
-    if not defaults:
-        defaults = []
-    requireds = [index < len(arguments) - len(defaults)  for index in range(len(arguments)) ]
-    for i in range( len(arguments) - 1, -1, -1) :
-        fn = click.argument(arguments[i], required=requireds[i])(fn)
-    fn = cli.command()(fn)
-    return fn
 
-@magic
 def collection_create(name, description=''):
     body = swagger_client.MainCreateCollectionBody(name=name, description=description)
     return collectionsApi.collections_post(body)
 
 
 #is this deprecated? I don't see it in the docs
-@magic
 def shuttle_create():
     response = requests.post(ESTUARY_URL+'admin/shuttle/init',
                   headers={'Authorization': 'Bearer '+ESTUARY_KEY})
     return response.json()
 
 
-@magic
 def autoretrieve_create(pub_key, addresses):
     return autoretrieveApi.admin_autoretrieve_init_post(addresses, pub_key)
 
-@magic
 def autoretrieve_list():
     return autoretrieveApi.admin_autoretrieve_list_get()
 
 
-@magic
 def autoretrieve_heartbeat(token):
     return autoretrieveApi.autoretrieve_heartbeat_post(token)
 
@@ -135,7 +103,6 @@ def _add_dir(path, collection_uuid='', root_collection_path=''):
 
 
 
-@magic
 def content_add(path, create_collection=False):
     if os.path.isfile(path):
         return _add_file(path)
@@ -155,7 +122,6 @@ def content_add(path, create_collection=False):
 
     return responses, collection
 
-@magic
 def collection_list():
     return collectionsApi.collections_get()
 
@@ -163,7 +129,6 @@ def collection_list():
 
 #TODO I don't really understand this method 
 # I think we want to use this https://github.com/snissn/estuary-swagger-clients/blob/main/python/docs/CollectionsApi.md#collections_coluuid_get
-@magic
 def collection_fs_list(collection_uuid, collection_path, recursive=False):
     query_params = f'col={collection_uuid}&dir={collection_path}'
 
@@ -184,34 +149,28 @@ def collection_fs_list(collection_uuid, collection_path, recursive=False):
 
 
 #todo what is recursive and how should it be used?
-@magic
 def collection_list_content(collection_uuid, collection_path='', recursive=False):
     return collectionsApi.collections_coluuid_get(collection_uuid, dir=collection_path)
 
 
-@magic
 def collection_commit(collection_uuid):
     return collectionsApi.collections_coluuid_commit_post(collection_uuid)
 
-@magic
 def content_add_ipfs(ipfs):
     body = swagger_client.UtilContentAddIpfsBody(root=ipfs)
     return contentApi.content_add_ipfs_post(body)
 
 # lists all contents for this user
-@magic
 def content_list():
     limit = '0' # seems to be ignored #TODO what is limit
     return collectionsApi.content_stats_get(limit)
     #TODO old version sorted by id, do we need that?
 
 # creates a new pin
-@magic
 def pin_create(cid, name):
     return pinningApi.pinning_pins_post(cid, name)
 
 # list all pins for this user
-@magic
 def pin_list():
     return pinningApi.pinning_pins_get()
